Refactor and clean up codebase, remove deprecated and unused code
This commit removes deprecated methods and structs, eliminates unused test-only constructors, and simplifies logic in several modules. It also updates error handling, improves code clarity, and removes or consolidates redundant code and comments across the backend, CLI, and link modules. Minor documentation and comment updates are included, as well as small bug fixes and code style improvements.

+++ backend/crates/kalamdb-filestore/src/object_store_ops.rs
@@ -250,7 +250,7 @@ pub async fn delete_prefix(
 
     while let Some(result) = stream.next().await {
         let meta = result.map_err(|e| FilestoreError::ObjectStore(e.to_string()))?;
-        total_bytes += meta.size as u64;
+        total_bytes += meta.size;
         paths_to_delete.push(meta.location);
     }
